+++ src/clojure_lsp/refactor/transform.clj
@@ -1030,3 +1030,20 @@
 
             ;; No source paths besides current one
             :else nil))))))
+
+(defn suppress-diagnostic [zloc diagnostic-code]
+  (let [diagnostic-code (if (= "unused-public-var" diagnostic-code)
+                          "clojure-lsp/unused-public-var"
+                          diagnostic-code)
+        form-zloc (z/up (edit/find-op zloc))
+        {form-row :row form-col :col :as form-pos} (-> form-zloc z/node meta)
+        loc-w-comment (z/edit-> form-zloc
+                                (z/insert-left (n/uneval-node (cond-> [(n/map-node [(n/keyword-node :clj-kondo/ignore)
+                                                                                    (n/spaces 1)
+                                                                                    (n/vector-node [(keyword diagnostic-code)])])
+                                                                       (n/newlines 1)]
+                                                                (> (dec form-col) 0) (conj (n/spaces (dec form-col)))))))]
+    [{:loc loc-w-comment
+      :range (assoc form-pos
+                    :end-row form-row
+                    :end-col form-col)}]))