+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Fix expand-let bug that occurs when a list form precedes let. #590
   - Add new command to create test for function at point. #582
   - Add `private` to documentSymbol to make clear that a var or function is  `private`.
+  - Add new code action `Suppress xxx diagnostic`, adding clj-kondo comment code to ignore the diagnostic. #591
 
 ## 2021.09.30-15.28.01